requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
gunicorn==21.2.0
sqlalchemy==2.0.44
pymysql==1.1.2
python-dotenv==1.0.0
python-multipart==0.0.6